requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
alembic==1.13.3
pydantic[email]==2.9.2
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
websockets==13.1
aiosqlite==0.20.0
pydantic_settings==2.5.2
jinja2==3.1.4
greenlet==3.1.1